Fixed some issuess

+++ Sand!/src/globals.cpp
@@ -5,5 +5,5 @@ std::list<GameObjects*> gameObject;
 int windowWidth = 1080;
 int windowHeight = 720;
 
-float mouseX = GetMouseX();
-float mouseY = GetMouseY();
+float mouseX = static_cast<float>(GetMouseX());
+float mouseY = static_cast<float>(GetMouseY());
